ATM
This is an application created via JavaScript that acts as an ATM machine. In this code, you will see how users are able to withdraw/deposit from the console starting off with an account balance of 0, and then how it is affected from further withdrawals/deposits. Once finished, the user types "done" before the application thanks them and returns back to its initial statement.